move pkg/api dependencies on util/subnet to api/util/subnet
