+++ pkg/api/v20220401/openshiftcluster_validatestatic.go
@@ -13,10 +13,10 @@ import (
 	"github.com/Azure/go-autorest/autorest/azure"
 
 	"github.com/Azure/ARO-RP/pkg/api"
+	api_subnet "github.com/Azure/ARO-RP/pkg/api/util/subnet"
 	"github.com/Azure/ARO-RP/pkg/api/validate"
 	"github.com/Azure/ARO-RP/pkg/util/immutable"
 	"github.com/Azure/ARO-RP/pkg/util/pullsecret"
-	"github.com/Azure/ARO-RP/pkg/util/subnet"
 	"github.com/Azure/ARO-RP/pkg/util/uuid"
 )
 
@@ -276,11 +276,11 @@ func (sv openShiftClusterStaticValidator) validateWorkerProfile(path string, wp 
 	default:
 		return api.NewCloudError(http.StatusBadRequest, api.CloudErrorCodeInvalidParameter, path+".encryptionAtHost", "The provided value '%s' is invalid.", wp.EncryptionAtHost)
 	}
-	workerVnetID, _, err := subnet.Split(wp.SubnetID)
+	workerVnetID, _, err := api_subnet.Split(wp.SubnetID)
 	if err != nil {
 		return err
 	}
-	masterVnetID, _, err := subnet.Split(mp.SubnetID)
+	masterVnetID, _, err := api_subnet.Split(mp.SubnetID)
 	if err != nil {
 		return err
 	}
