+++ pkg/util/subnet/subnet.go
@@ -5,9 +5,7 @@ package subnet
 
 import (
 	"context"
-	"fmt"
 	"net"
-	"strings"
 
 	mgmtnetwork "github.com/Azure/azure-sdk-for-go/services/network/mgmt/2020-08-01/network"
 	"github.com/Azure/go-autorest/autorest"
@@ -15,6 +13,7 @@ import (
 	"github.com/apparentlymart/go-cidr/cidr"
 
 	"github.com/Azure/ARO-RP/pkg/api"
+	api_subnet "github.com/Azure/ARO-RP/pkg/api/util/subnet"
 	"github.com/Azure/ARO-RP/pkg/util/azureclient"
 	"github.com/Azure/ARO-RP/pkg/util/azureclient/mgmt/network"
 )
@@ -50,7 +49,7 @@ func (m *manager) Get(ctx context.Context, subnetID string) (*mgmtnetwork.Subnet
 }
 
 func (m *manager) get(ctx context.Context, subnetID, expand string) (*mgmtnetwork.Subnet, error) {
-	vnetID, subnetName, err := Split(subnetID)
+	vnetID, subnetName, err := api_subnet.Split(subnetID)
 	if err != nil {
 		return nil, err
 	}
@@ -111,7 +110,7 @@ func (m *manager) GetHighestFreeIP(ctx context.Context, subnetID string) (string
 
 // CreateOrUpdate updates the linked subnet
 func (m *manager) CreateOrUpdate(ctx context.Context, subnetID string, subnet *mgmtnetwork.Subnet) error {
-	vnetID, subnetName, err := Split(subnetID)
+	vnetID, subnetName, err := api_subnet.Split(subnetID)
 	if err != nil {
 		return err
 	}
@@ -165,54 +164,4 @@ func (m *manager) createOrUpdateSubnets(ctx context.Context, subnets []*mgmtnetw
 		}
 	}
 	return nil
-}
-
-// Split splits the given subnetID into a vnetID and subnetName
-func Split(subnetID string) (string, string, error) {
-	parts := strings.Split(subnetID, "/")
-	if len(parts) != 11 {
-		return "", "", fmt.Errorf("subnet ID %q has incorrect length", subnetID)
-	}
-
-	return strings.Join(parts[:len(parts)-2], "/"), parts[len(parts)-1], nil
-}
-
-// NetworkSecurityGroupID returns the NetworkSecurityGroup ID for a given subnet
-// ID
-func NetworkSecurityGroupID(oc *api.OpenShiftCluster, subnetID string) (string, error) {
-	infraID := oc.Properties.InfraID
-	if infraID == "" {
-		infraID = "aro"
-	}
-	isWorkerSubnet := false
-	for _, s := range oc.Properties.WorkerProfiles {
-		if strings.EqualFold(subnetID, s.SubnetID) {
-			isWorkerSubnet = true
-			break
-		}
-	}
-	return NetworkSecurityGroupIDExpanded(oc.Properties.ArchitectureVersion, oc.Properties.ClusterProfile.ResourceGroupID, infraID, isWorkerSubnet)
-}
-
-// NetworkSecurityGroupIDExpanded returns the NetworkSecurityGroup ID for a given subnetID, without the OpenShift Cluster document
-func NetworkSecurityGroupIDExpanded(architectureVersion api.ArchitectureVersion, resourceGroupID, infraID string, isWorkerSubnet bool) (string, error) {
-	switch architectureVersion {
-	case api.ArchitectureVersionV1:
-		return networkSecurityGroupIDV1(resourceGroupID, infraID, isWorkerSubnet), nil
-	case api.ArchitectureVersionV2:
-		return networkSecurityGroupIDV2(resourceGroupID, infraID), nil
-	default:
-		return "", fmt.Errorf("unknown architecture version %d", architectureVersion)
-	}
-}
-
-func networkSecurityGroupIDV1(resourceGroupID, infraID string, isWorkerSubnet bool) string {
-	if isWorkerSubnet {
-		return resourceGroupID + "/providers/Microsoft.Network/networkSecurityGroups/" + infraID + NSGNodeSuffixV1
-	}
-	return resourceGroupID + "/providers/Microsoft.Network/networkSecurityGroups/" + infraID + NSGControlPlaneSuffixV1
-}
-
-func networkSecurityGroupIDV2(resourceGroupID, infraID string) string {
-	return resourceGroupID + "/providers/Microsoft.Network/networkSecurityGroups/" + infraID + NSGSuffixV2
 }
